pyproject.toml
[project]
name = "battery-dispatch-rl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "cvxpy[clarabel]>=1.8.1",
    "entsoe-py>=0.7.10",
    "gymnasium>=1.2.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "stable-baselines3[extra]>=2.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
bess-dispatch = "bess_dispatch.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bess_dispatch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
